refactor: organize tools into separate files by functionality
- Created server/tools/ui.ts for UI interaction tools (trigger_action, risky_eval, emulate_clicks, fill_dialog)
- Created server/tools/texture.ts for texture management tools (create_texture, apply_texture, add_texture_group, list_textures, get_texture)
- Created server/tools/element.ts for element/outliner tools (remove_element, add_group, list_outline, duplicate_element, rename_element)
- Created server/tools/project.ts for project management tools (create_project)
- Created server/tools/camera.ts for camera and screenshot tools (capture_screenshot, capture_app_screenshot, set_camera_angle)
- Created server/tools/import.ts for import/export tools (from_geo_json)
- Moved knife_tool to server/tools/mesh.ts
- Moved remaining paint tools to server/tools/paint.ts (paint_with_brush, create_brush_preset, load_brush_preset, texture_selection, texture_layer_management)
- Updated server/tools.ts to only contain imports

This refactor improves code organization and maintainability by grouping related tools together based on their functionality.
